requirements.txt
appdirs==1.4.2
blessings==1.6
bpython==0.16
coverage==4.3.4
curtsies==0.2.11
dj-database-url==0.4.2
dj-email-url==0.0.10
Django==1.10.6
django-assets==0.12
django-debug-toolbar==1.6
django-filter==1.0.1
djangorestframework==3.5.4
factory-boy==2.8.1
Faker==0.7.9
greenlet==0.4.12
gunicorn==19.6.0
invoke==0.15.0
packaging==16.8
psycopg2==2.7
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
requests==2.13.0
six==1.10.0
sqlparse==0.2.3
wcwidth==0.1.7
webassets==0.12.1
whitenoise==3.3.0
